requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
eth-account>=0.9.0